Deprecate busted functionality, fix priority inversion
Attempt to skip the postmaster and the logger in addition to the
archiver when sleeping.

Deprecate the renicing targets, because it wasn't being used and would
require refactoring to support the new definition of processes exempt
from throttling.

Remove test output summary, because it was always broken, but it
doesn't seem important right now to fix it.

Signed-off-by: Daniel Farina <daniel@fdr.io>