Resolve contradictory handling of postmaster nap exemption
Now, postmaster (or whatever process is selected by `--pid` or
`--port`) is *not* included among the napping processes.  Postmaster
is the postgres process supervisor and, if practical, it would be
better to not interfere with it.

To allow `ratchet self-test` to demonstrate `ratchet`'s functionality,
it needed complications to mimic the postmaster/backend topology of
processes.

In principle, exempting postmaster could make `ratchet` not effective
on databases with extremely high process churn, although those
processes would somehow have to generate a meaningful amount of WAL
before exiting and churning again.  It's not clear to me that is a
case practical to worry about.

Looking at it, this program had internal contradictions: in one place,
it purposefully wanted to include postmaster, in a comment: `# Stop
the parent first, so it cannot create children.`

But, in another there was code to exempt it: `is_postmaster =
contains('-D') and contains('config_file=')`.  However, that predicate
was never true, because `ps --ppid` does not return the process
information of the PID targeted by `--ppid`.

It also had an unused data structure: `children_maybe_stopped`.